actualizacion

+++ Proyecto_BooBookStore/src/main/java/com/cibertec/springboot/web/app/controllers/LibroController.java
@@ -8,9 +8,8 @@ import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
@@ -31,7 +30,9 @@ import com.cibertec.springboot.web.app.models.entity.Libro;
 import com.cibertec.springboot.web.app.models.service.ICategoriaService;
 import com.cibertec.springboot.web.app.models.service.ILibroService;
 import com.cibertec.springboot.web.app.models.service.IUploadFileService;
+import com.cibertec.springboot.web.app.util.constants.Constant;
 import com.cibertec.springboot.web.app.util.paginator.PageRender;
+import com.google.gson.Gson;
 
 @Controller
 @SessionAttributes(names = {"libro", "empleado"})
@@ -64,15 +65,13 @@ public class LibroController {
 	@RequestMapping(value="/listado", method = RequestMethod.GET)
 	public String listado(@RequestParam(name="page", defaultValue = "0") 
 	int page, Model modeloListado, Authentication authentication) {
-		Pageable pageRequest = PageRequest.of(page, 5);
-		Page<Libro> libros = libroService.findAll(pageRequest);
+		Page<Libro> libros = libroService.findAll(page, Constant.PAGE_SIZE);
 		PageRender<Libro> pageRender = new PageRender<>("/libro/listado", libros);
-				
+		
 		modeloListado.addAttribute("titulo", "Listado de Libros");
 		modeloListado.addAttribute("libros", libros);
 		modeloListado.addAttribute("page", pageRender);
 		
-		System.out.println(modeloListado.getAttribute("empleado"));
 		return "libro/listado";
 	}
 	
@@ -96,13 +95,12 @@ public class LibroController {
 		if (!foto.isEmpty()) {
 			
 			if (libro.getId() != null && libro.getId() > 0 &&
-					libro.getFoto() != null && libro.getFoto().length()> 0) {				
-				uploadFileService.delete(libro.getFoto());			
-			}						
-			String uniqueFilename = null;						
+					libro.getFoto() != null && libro.getFoto().length()> 0) {
+				uploadFileService.delete(libro.getFoto());
+			}
+			String uniqueFilename = null;
 			try {
 				uniqueFilename = uploadFileService.copy(foto);
-				
 			} catch (IOException e) {
 				e.printStackTrace();
 			}		
@@ -110,31 +108,61 @@ public class LibroController {
 			attributes.addFlashAttribute("info", "Has subido correctamente '" + uniqueFilename + "'");
 			libro.setFoto(uniqueFilename);
 		}
-		
-		libroService.save(libro);
+		ResponseEntity<String> response = new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
+		if (libro.getId() != null) {
+			response = libroService.update(libro);
+		} else {
+			response = libroService.save(libro);			
+		}
+		if (response.getStatusCode() == HttpStatus.OK || response.getStatusCode() == HttpStatus.CREATED) {
+		    attributes.addFlashAttribute("success", "Libro guardado correctamente.");
+		} else {
+		    attributes.addFlashAttribute("error", "Ocurrió un error al guardar.");
+		}
 		status.setComplete();
-		attributes.addFlashAttribute("success", "Libro guardado correctamente.");
 		return "redirect:/libro/listado";
 	}
 	
 	@RequestMapping(value = "/editar/{id}")
-	public String editar(Model model, @PathVariable Long id) {
+	public String editar(Model model, @PathVariable Long id, RedirectAttributes attributes) {
+		Libro libro = null;
+		ResponseEntity<String> response = libroService.findOne(id);
+		if (response.getStatusCode() == HttpStatus.OK) {
+			Gson gson = new Gson();
+			libro = gson.fromJson(response.getBody(), Libro.class);
+			model.addAttribute("titulo","Formulario de Libro");
+			model.addAttribute("libro", libro);
+			model.addAttribute("categorias",categoriaService.findAll());
+		} else {
+			attributes.addFlashAttribute("error", "No se encontró un libro con el id indicado.");
+			return "redirect:/libro/listado";
+		}
 		
-		model.addAttribute("titulo","Formulario de Libro");
-		model.addAttribute("libro", libroService.findOne(id));
-		model.addAttribute("categorias",categoriaService.findAll());
 		return "libro/registro";
 	}
 	
 	@RequestMapping(value = "/eliminar/{id}")
 	public String eliminar(@PathVariable Long id, RedirectAttributes attributes) {
-		Libro libro = libroService.findOne(id);
-		libroService.delete(id);
-		attributes.addFlashAttribute("success", "Libro eliminado correctamente.");
-		if (!(libro.getFoto() == null)) {
-			if (uploadFileService.delete(libro.getFoto())) {
-				attributes.addAttribute("info", "Foto " + libro.getFoto() + " eliminada con éxito.");			
+		Libro libro = null;
+		ResponseEntity<String> response = libroService.findOne(id);
+		if (response.getStatusCode() == HttpStatus.OK) {
+			Gson gson = new Gson();
+			libro = gson.fromJson(response.getBody(), Libro.class);
+		} else {
+			attributes.addFlashAttribute("error", "No se encontró un libro con el id indicado.");
+			return "redirect:/libro/listado";
+		}
+		
+		response = libroService.delete(id);
+		if (response.getStatusCode() == HttpStatus.OK) {
+			attributes.addFlashAttribute("success", "Libro eliminado correctamente.");
+			if (libro != null && libro.getFoto() != null) {
+				if (uploadFileService.delete(libro.getFoto())) {
+					attributes.addFlashAttribute("info", "Foto " + libro.getFoto() + " eliminada con éxito.");
+				}
 			}
+		} else {
+		    attributes.addFlashAttribute("error", "Ocurrió un error al eliminar.");
 		}
 		return "redirect:/libro/listado";
 	}
